# api-docs/api-docs.yaml
openapi: 3.1.0
info:
  title: LibroRed API
  description: REST API for LibroRed - Web Application for Book Lending Between Individuals
  contact:
    name: Team 13
    url: https://github.com/medinaymedia
    email: am.juradoc@alumnos.urjc.es
  license:
    name: Apache 2.0
    url: http://springdoc.org
  version: '1.0'
servers:
- url: https://localhost:8443
  description: Development server
...
paths:
  /api/v1/users:
    get:
      tags:
      - Users
      summary: Get all users
      description: Retrieve a list of all users
      operationId: getAllUsers
      responses:
        '200':
          description: Users retrieved successfully
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/UserDTO'
    post:
      tags:
      - Users
      summary: Create a new user
      description: Create a new user account
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        '201':
          description: User created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid user data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        '500':
          description: Internal server error
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
  /api/v1/users/{id}:
    get:
      tags:
      - Users
      summary: Get user by ID
      description: Retrieve a specific user by their ID
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: User found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
    put:
      tags:
      - Users
      summary: Update user
      description: Update an existing user
      operationId: updateUser
      parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDTO'
        required: true
      responses:
        '200':
          description: User updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        '404':
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
        '400':
          description: Invalid user data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UserDTO'
    delete:
      tags:
      - Users
      summary: Delete user
      description: Delete a user by ID
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: User ID
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '204':
          description: User deleted successfully
        '404':
          description: User not found
...
  /api/users/verify-password:
    post:
      tags:
      - User Management
      summary: Verify current password
      description: Verify user's current password for account changes
      operationId: verifyPassword
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
              required:
                - currentPassword
      responses:
        '200':
          description: Password verification result
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: User not authenticated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/users/update-username:
    post:
      tags:
      - User Management
      summary: Update username
      description: Update authenticated user's username
      operationId: updateUsername
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                newUsername:
                  type: string
              required:
                - newUsername
      responses:
        '200':
          description: Username update result
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: User not authenticated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /api/users/update-password:
    post:
      tags:
      - User Management
      summary: Update password
      description: Update authenticated user's password
      operationId: updatePassword
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                currentPassword:
                  type: string
                newPassword:
                  type: string
              required:
                - currentPassword
                - newPassword
      responses:
        '200':
          description: Password update result
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: User not authenticated
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
components:
  schemas:
    UserDTO:
      type: object
      description: User data
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum:
          - ROLE_USER
          - ROLE_ADMIN
...
    ApiResponse:
      type: object
      description: Generic API response
      properties:
        success:
          type: boolean
        message:
          type: string
        error:
          type: string
...